+++ changelog.md
@@ -2,6 +2,28 @@
 
 > All Time Stamps: GMT+8
 
+## 0.3.0.2 _2020-05-21 17:00_
+
+* 加入方块和物品 added blocks and items
+  * 硬实泥土矿 solid dirt ore block
+  * 硬实沙矿 solid sand ore block
+  * 硬实石头矿 solid stone ore block
+  * 叶绿粉 chloroplast dust item
+  * 彩色玻璃块 colorful glass block
+  * 激流推进器 torrential thruster
+  * 导管 tube item
+  * 一次性空气罐 disposable air tank item
+  * 空气罐 air tank item
+  * 空气泵 air pump block
+  * 空气弹 air bomb item
+* 加入特性 added traits
+  * 导管升级(护甲) tube upgraded(armor)
+  * 激流 torrential
+  * 水合(护甲) hydrating(armor)
+* 加入状态效果 added potion effects
+  * 发泡 bubbling
+  * 乱流 turbulent
+
 ## 0.3.0.1 _2020-05-20 18:00_
 
 * 重写世界生成 rewrote world generation